Repackage git and svn into sub-packages
