+++ plugin/versioning/src/test/kotlin/net/twisterrob/gradle/android/AndroidVersionPluginIntgTest.kt
@@ -7,11 +7,12 @@ import net.twisterrob.gradle.test.assertNoOutputLine
 import net.twisterrob.gradle.test.assertSuccess
 import net.twisterrob.gradle.test.root
 import net.twisterrob.gradle.vcs.createTestFileToCommit
-import net.twisterrob.gradle.vcs.doCheckout
-import net.twisterrob.gradle.vcs.doCommitSingleFile
-import net.twisterrob.gradle.vcs.doCreateRepository
-import net.twisterrob.gradle.vcs.git
-import net.twisterrob.gradle.vcs.svn
+import net.twisterrob.gradle.vcs.git.doCommitSingleFile
+import net.twisterrob.gradle.vcs.git.git
+import net.twisterrob.gradle.vcs.svn.doCheckout
+import net.twisterrob.gradle.vcs.svn.doCommitSingleFile
+import net.twisterrob.gradle.vcs.svn.doCreateRepository
+import net.twisterrob.gradle.vcs.svn.svn
 import org.intellij.lang.annotations.Language
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
